+++ RELEASE_NOTES.rst
@@ -15,6 +15,7 @@ Vagga 0.4.1
 * Renamed and fixed ``vagga_network`` command as ``vagga _network`` subcommand
 * The pip cache is now namespaces over distro and version (was singleton)
 * Vagga now cleans apt lists cache on failed ``apt-get update`` (#108)
+* Add ``UbuntuPPA`` and ``AptTrust`` build steps
 
 Vagga 0.4.0
 ===========